requirements.txt
requests
beautifulsoup4
psycopg2-binary
feedparser
python-dotenv
